Adding ability to ship/configure the TaskWorker tarball
Previously, the tarball was hardcoded to a specific url. This
commit lets you either change the url with:

config.Debug.taskManagerRunTarballLocation

OR, you can set the url to 'local' and set the location of your
source trees with:

config.Debug.taskManagerCodeLocation

In the latter case, on submission a tarball will be created and then
passed along with the condor file-transfer mechanism.